fix: Add CSR/CSC validation to diagnose mul64 UMFPACK invalid matrix on CI
- Add ValidateCsr() in umfpack_ffi.cpp: validates both CSR input and CSC
  output before passing to umfpack_di_symbolic, with detailed error messages
  (which row, which indices) to diagnose the CI-only failure
- Add Python-side CSR validation in full_mna.py at setup time (vectorized)
- Fix int32 overflow in build_csr_arrays (sparse.py:105): cast to int64
  before multiplication to prevent overflow for large matrices (666k x 666k)
- Include matrix dimensions (n, nnz) in all UMFPACK error messages

Co-developed-by: Claude Code v2.1.44 (claude-opus-4-6)

+++ vajax/analysis/sparse.py
@@ -102,7 +102,8 @@ def build_csr_arrays(
     n_rows = shape[0]
 
     # Create linear indices for sorting (row-major order)
-    linear_idx = rows * shape[1] + cols
+    # Use int64 to avoid overflow for large matrices (e.g., 666k x 666k)
+    linear_idx = rows.astype(jnp.int64) * shape[1] + cols.astype(jnp.int64)
 
     # Sort by linear index to group duplicates
     sort_order = jnp.argsort(linear_idx)
